+++ plugins/lookup/kv.py
@@ -20,6 +20,8 @@ options:
 extends_documentation_fragment:
   - flowerysong.hvault.base
   - flowerysong.hvault.base.PLUGINS
+  - flowerysong.hvault.auth_token
+  - flowerysong.hvault.auth_token.PLUGINS
 """
 
 EXAMPLES = """
@@ -41,8 +43,9 @@ from ..plugin_utils.lookup import HVaultLookupBase
 
 class LookupModule(HVaultLookupBase):
     def run(self, terms, variables=None, **kwargs):
-        self.set_options(direct=kwargs)
+        self.init_options(variables=variables, direct=kwargs)
         self.set_option('raw', False)
+        self.config_client()
 
         mount = self.get_option('mount_point')
         terms = ['/'.join((mount, x)) for x in terms]