+++ plugins/lookup/kv_v2.py
@@ -25,6 +25,8 @@ options:
 extends_documentation_fragment:
   - flowerysong.hvault.base
   - flowerysong.hvault.base.PLUGINS
+  - flowerysong.hvault.auth_token
+  - flowerysong.hvault.auth_token.PLUGINS
 """
 
 EXAMPLES = """
@@ -50,8 +52,9 @@ from ..plugin_utils.lookup import HVaultLookupBase
 
 class LookupModule(HVaultLookupBase):
     def run(self, terms, variables=None, **kwargs):
-        self.set_options(direct=kwargs)
+        self.init_options(variables=variables, direct=kwargs)
         self.set_option('raw', False)
+        self.config_client()
 
         mount = self.get_option('mount_point')
         version = self.get_option('version')