+++ plugins/lookup/read.py
@@ -19,6 +19,8 @@ options:
 extends_documentation_fragment:
   - flowerysong.hvault.base
   - flowerysong.hvault.base.PLUGINS
+  - flowerysong.hvault.auth_token
+  - flowerysong.hvault.auth_token.PLUGINS
 """
 
 EXAMPLES = """
@@ -60,6 +62,7 @@ from ..plugin_utils.lookup import HVaultLookupBase
 
 class LookupModule(HVaultLookupBase):
     def run(self, terms, variables=None, **kwargs):
-        self.set_options(direct=kwargs)
+        self.init_options(variables=variables, direct=kwargs)
+        self.config_client()
 
         return super().run(terms)