+++ plugins/modules/aws_auth_role.py
@@ -17,6 +17,7 @@ description:
 version_added: 0.2.0
 extends_documentation_fragment:
   - flowerysong.hvault.base
+  - flowerysong.hvault.auth_token
   - flowerysong.hvault.role
   - flowerysong.hvault.token
 options: