+++ plugins/modules/ldap_group.py
@@ -16,6 +16,7 @@ description:
 version_added: 0.2.0
 extends_documentation_fragment:
   - flowerysong.hvault.base
+  - flowerysong.hvault.auth_token
 options:
   mount_point:
     description: